Fix oic.sec.ace.json reference issues.
Use a definition to reference instead of the whole schema file.
There seem to be an issue with the validator when referencing the whole
schema file instead of a single definition. I suspect this is a similar
issue I have seen with other validators that nicely bundle referenced,
adapt the local references to global references in the bundle, but fail
to strip the id from the externally imported files.
The result of this is that the global bundle reference gets resolved
against the id of the imported file instead of the id of the global
bundle.
This results in a similar side effect as observed with atom api-workbench
validator.
Subsequent fixes will be needed to clean up all internal schema's to use
definitions rather than file references.